move Province to its own entity
