+++ packages/types/src/entities/ConsentCategory.ts
@@ -10,14 +10,3 @@ const CONSENT_CATEGORIES = [
 
 export const ConsentCategory = z.enum(CONSENT_CATEGORIES);
 export type ConsentCategory = z.infer<typeof ConsentCategory>;
-
-const CONSENT_GROUPS = [
-	'ADULT_CONSENT',
-	'ADULT_CONSENT_SUBSTITUTE_DECISION_MAKER',
-	'GUARDIAN_CONSENT_OF_MINOR',
-	'GUARDIAN_CONSENT_OF_MINOR_INCLUDING_ASSENT',
-	'YOUNG_ADULT_CONSENT',
-] as const;
-
-export const ConsentGroup = z.enum(CONSENT_GROUPS);
-export type ConsentGroup = z.infer<typeof ConsentGroup>;
