+++ packages/types/src/entities/index.ts
@@ -1,2 +1,6 @@
 export * from './User.js';
 export * from './ConsentCategory.js';
+export * from './ConsentGroup.js';
+export * from './ConsentQuestion.js';
+export * from './ParticipantIdentification.js';
+export * from './Province.js';
